pyproject.toml
[project]
name = "cin-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=0.1.0",
    "hello-agents==0.1.1",
    "typing>=3.10.0.0",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
]
